// src/app/models/usuario.model.ts
export class Usuario {

    idUsuario?: number;
    nombres?: string;
    apaterno?: string;
    amaterno?: string;
    fechaNacimiento?: string;
    fechaRegistro?: string;
    email?: string;
    estado?: string;
    celular?: string;
    direccion?: string;
    login?: string;
    password?: string;
    dni?: string;
    idTipoUsuario?: number;
    idUbigeo?: number;
}